backend_restaurant/requirements.txt
asgiref==3.8.1
django==5.0.6
djangorestframework==3.15.1
drf-yasg==1.21.7
sqlparse==0.5.0
django-jazzmin==2.6.0
requests